requirements.txt
asgiref==3.5.0
asttokens==2.0.5
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
backcall==0.2.0
backports.entry-points-selectable==1.1.1
binance-futures-connector==3.2.0
black==22.10.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.4
constantly==15.1.0
cryptography==38.0.4
debugpy==1.5.1
decorator==5.1.1
distlib==0.3.4
Django==4.0.1
django-channels==0.7.0
django-ckeditor==6.5.1
django-filter==22.1
django-js-asset==2.0.0
djangorestframework==3.14.0
entrypoints==0.3
executing==0.8.2
filelock==3.4.0
Flask==2.1.2
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipykernel==6.7.0
ipython==8.0.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
mysqlclient==2.1.0
nest-asyncio==1.5.4
numpy==1.22.4
oauthlib==3.2.2
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.0
prompt-toolkit==3.0.24
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
pyOpenSSL==22.1.0
python-dateutil==2.8.2
pytz==2022.4
pywin32==303
pyzmq==22.3.0
requests==2.28.1
requests-oauthlib==1.3.1
scikit-learn==1.1.1
scipy==1.8.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.4
threadpoolctl==3.1.0
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.13
virtualenv==20.10.0
virtualenvwrapper-win==1.2.7
wcwidth==0.2.5
websocket-client==1.4.2
Werkzeug==2.1.2
zope.interface==5.5.2
